First e2e crack at updated workflow
