+++ yuptoo/config/base.py
@@ -23,3 +23,5 @@ UPLOAD_TOPIC = os.getenv('UPLOAD_TOPIC', 'platform.inventory.host-ingress')
 KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE = os.getenv(
     'KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE', 2097152
 )
+HOSTS_TRANSFORMATION_ENABLED = os.getenv('HOSTS_TRANSFORMATION_ENABLED', default=True)
+HOSTS_UPLOAD_FUTURES_COUNT = os.getenv('HOSTS_UPLOAD_FUTURES_COUNT', default=100)
