+++ yuptoo/processor/report_slice_processor.py
@@ -1,93 +1,50 @@
-#
-# Copyright 2019 Red Hat, Inc.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-"""Report Slice Processor."""
-
-import asyncio
 import base64
 import copy
 import json
-import logging
-from processor.report_consumer import (KAFKA_ERRORS,
-                                       KafkaMsgHandlerError,
-                                       QPCReportException)
+import re
+from confluent_kafka import Producer, KafkaException
 
-from config.base import (
+from yuptoo.processor.exceptions import KafkaMsgHandlerError
+from yuptoo.processor.host_transform_processor import HostTransformProcessor
+from yuptoo.config.base import (
     HOSTS_TRANSFORMATION_ENABLED, KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE,
-    UPLOAD_TOPIC
+    UPLOAD_TOPIC, INSIGHTS_KAFKA_ADDRESS, HOSTS_UPLOAD_FUTURES_COUNT, get_logger
 )
 
-from yuptoo.processor.host_transformation import (
-    transform_tags, remove_display_name, remove_empty_ip_addresses,
-    transform_mac_addresses, remove_invalid_bios_uuid,
-    transform_os_release, transform_os_kernel_version,
-    transform_network_interfaces, remove_installed_packages,
-    TRANSFORMED_DICT
-)
-
-LOG = logging.getLogger(__name__)
+LOG = get_logger(__name__)
+TRANSFORMED_DICT = dict({'removed': [], 'modified': [], 'missing_data': []})
 
 
-class RetryUploadTimeException(Exception):
-    """Use to report upload errors that should be retried on time."""
-
-    pass
-
-
-class RetryUploadCommitException(Exception):
-    """Use to report upload errors that should be retried on commit."""
-
-    pass
-
-
-class ReportSliceProcessor:  # pylint: disable=too-many-instance-attributes
+class ReportSliceProcessor(HostTransformProcessor):  # pylint: disable=too-many-instance-attributes
     """Class for processing report slices that have been created."""
 
-    def __init__(self):
-        self.producer = AIOKafkaProducer(
-            loop=SLICE_PROCESSING_LOOP, bootstrap_servers=INSIGHTS_KAFKA_ADDRESS,
-            max_request_size=KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE
-        )
-
-    def pass_to_report_slice_processor(self, options):
-        pass
+    def __init__(self, report_platform_id):
+        self.report_platform_id = report_platform_id
 
     def transform_single_host(self, request_id, host_id, host: dict):
         """Transform 'system_profile' fields."""
         transformed_obj = copy.deepcopy(TRANSFORMED_DICT)
         if 'system_profile' in host:
-            host, transformed_obj = transform_os_release(
+            host, transformed_obj = self.transform_os_release(
                 host, transformed_obj)
-            host, transformed_obj = transform_os_kernel_version(
+            host, transformed_obj = self.transform_os_kernel_version(
                 host, transformed_obj)
-            host, transformed_obj = transform_network_interfaces(
+            host, transformed_obj = self.transform_network_interfaces(
                 host, transformed_obj)
 
-        host, transformed_obj = remove_empty_ip_addresses(
+        host, transformed_obj = self.remove_empty_ip_addresses(
             host, transformed_obj)
-        host, transformed_obj = transform_mac_addresses(
+        host, transformed_obj = self.transform_mac_addresses(
             host, transformed_obj)
-        host, transformed_obj = remove_display_name(
+        host, transformed_obj = self.remove_display_name(
             host, transformed_obj)
-        host, transformed_obj = remove_invalid_bios_uuid(
+        host, transformed_obj = self.remove_invalid_bios_uuid(
             host, transformed_obj)
-        host, transformed_obj = transform_tags(host, transformed_obj)
+        host, transformed_obj = self.transform_tags(host, transformed_obj)
+
         host_request_size = bytes(json.dumps(host), 'utf-8')
         if len(host_request_size) >= KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE:
-            host, transformed_obj = remove_installed_packages(
+            host, transformed_obj = self.remove_installed_packages(
                 host, transformed_obj)
 
         self.print_transformed_info(request_id, host_id, transformed_obj)
@@ -95,6 +52,7 @@ class ReportSliceProcessor:  # pylint: disable=too-many-instance-attributes
 
     def print_transformed_info(self, request_id, host_id, transformed_obj):
         """Print transformed logs."""
+        prefix = 'Printing Transformed Logs'
         if transformed_obj is None:
             return
 
@@ -105,69 +63,82 @@ class ReportSliceProcessor:  # pylint: disable=too-many-instance-attributes
 
         if log_sections:
             log_message = (
-                'Transformed details host with id %s (request_id: %s):\n'
-                % (host_id, request_id)
+                '%s - Transformed details host with id %s (request_id: %s) for account=%s and report_platform_id=%s.'
             )
             log_message += '\n'.join(log_sections)
             LOG.info(
-                    prefix,
-                    log_message,
-                    account_number=self.account_number,
-                    report_platform_id=self.report_platform_id
+                log_message, prefix, host_id, request_id, self.account, self.report_platform_id
             )
+
+    def delivery_report(self, err, msg=None):
+        prefix = 'PUBLISH TO INVENTORY TOPIC ON KAFKA'
+
+        if err is not None:
+            LOG.error(
+                "Message delivery for topic %s failed for request_id [%s]: %s",
+                msg.topic(),
+                err,
+                self.request_id,
+            )
+        else:
+            LOG.info(
+                "%s - Message delivered to %s [%s] for request_id [%s]",
+                prefix,
+                msg.topic(),
+                msg.partition(),
+                self.request_id
+            )
+
+    def send_message(self, topic, msg):
+        try:
+
+            self.producer = Producer({
+                'bootstrap.servers': INSIGHTS_KAFKA_ADDRESS,
+                'message.max.bytes': KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE
+            })
+
+            bytes = json.dumps(msg, ensure_ascii=False).encode("utf-8")
+
+            self.producer.produce(topic, bytes, callback=self.delivery_report)
+            self.producer.poll(1)
+        except KafkaException:
+            LOG.exception(
+                "Failed to produce message to [%s] topic: %s", topic, self.request_id
+            )
+        finally:
+            self.producer.flush()
 
     # pylint:disable=too-many-locals
     # pylint: disable=too-many-statements
-    def upload_to_host_inventory_via_kafka(self, hosts):
-        """
-        Upload to the host inventory via kafka.
-
-        :param: hosts <list> the hosts to upload.
-        """
-        self.prefix = 'UPLOAD TO INVENTORY VIA KAFKA'
-        await self.producer.stop()
-        self.producer = AIOKafkaProducer(
-            loop=SLICE_PROCESSING_LOOP, bootstrap_servers=INSIGHTS_KAFKA_ADDRESS,
-            max_request_size=KAFKA_PRODUCER_OVERRIDE_MAX_REQUEST_SIZE
-        )
-        try:
-            await self.producer.start()
-        except (KafkaConnectionError, TimeoutError):
-            KAFKA_ERRORS.inc()
-            self.should_run = False
-            print_error_loop_event()
-            raise KafkaMsgHandlerError(
-                format_message(
-                    self.prefix,
-                    'Unable to connect to kafka server.',
-                    account_number=self.account_number,
-                    report_platform_id=self.report_platform_id))
+    def upload_to_host_inventory_via_kafka(self, hosts, b64_identity, request_id):
+        self.request_id = request_id
+        prefix = 'UPLOAD TO INVENTORY VIA KAFKA'
         total_hosts = len(hosts)
         count = 0
-        send_futures = []
-        associated_msg = []
-        report = self.report_or_slice.report
         cert_cn = None
+
         try:
-            b64_identity = json.loads(report.upload_srv_kafka_msg)['b64_identity']
             raw_b64_identity = base64.b64decode(b64_identity).decode('utf-8')
             identity = json.loads(raw_b64_identity)
+            self.account = identity['identity']['account_number']
             cert_cn = identity['identity']['system']['cn']
         except KeyError as err:
             LOG.error(
                 prefix, 'Invalid identity. Key not found: %s', err)
 
-        unique_id_base = '{}:{}:{}:'.format(report.request_id,
-                                            report.report_platform_id,
-                                            self.report_or_slice.report_slice_id)
+        unique_id_base = '{}:{}:'.format(self.request_id,
+                                         self.report_platform_id)
         try:  # pylint: disable=too-many-nested-blocks
             for host_id, host in hosts.items():
                 if HOSTS_TRANSFORMATION_ENABLED:
-                    host = self._transform_single_host(
-                        report.request_id, host_id, host)
+                    host = self.transform_single_host(
+                        self.request_id, host_id, host)
                     if cert_cn and ('system_profile' in host):
                         host['system_profile']['owner_id'] = cert_cn
-                system_unique_id = unique_id_base + host_id
+
+                report_slice_id = host.get('report_slice_id')
+                system_unique_id = unique_id_base + report_slice_id + host_id
+
                 count += 1
                 upload_msg = {
                     'operation': 'add_host',
@@ -175,42 +146,17 @@ class ReportSliceProcessor:  # pylint: disable=too-many-instance-attributes
                     'platform_metadata': {'request_id': system_unique_id,
                                           'b64_identity': b64_identity}
                 }
-                msg = bytes(json.dumps(upload_msg), 'utf-8')
-                future = await self.producer.send(UPLOAD_TOPIC, msg)
-                send_futures.append(future)
-                associated_msg.append(upload_msg)
+
+                self.send_message(UPLOAD_TOPIC, upload_msg)
                 if count % HOSTS_UPLOAD_FUTURES_COUNT == 0 or count == total_hosts:
                     LOG.info(
-                        format_message(
-                            self.prefix,
-                            'Sending %s/%s hosts to the inventory service.' % (count, total_hosts),
-                            account_number=self.account_number,
-                            report_platform_id=self.report_platform_id))
-                    try:
-                        await asyncio.wait(send_futures, timeout=HOSTS_UPLOAD_TIMEOUT)
-                        future_index = 0
-                        for future_res in send_futures:
-                            if future_res.exception():
-                                LOG.error(
-                                    'An exception occurred %s when trying to upload '
-                                    'the following message: %s',
-                                    future_res.exception(),
-                                    associated_msg[future_index])
-                            future_index += 1
-                    except Exception as error:  # pylint: disable=broad-except
-                        LOG.error('An exception occurred: %s', error)
-                    send_futures = []
+                        '%s - Sending %s/%s hosts to the inventory service for account=%s and report_platform_id=%s.',
+                        prefix, count, total_hosts, self.account, self.report_platform_id)
+
         except Exception as err:  # pylint: disable=broad-except
-            LOG.error(format_message(
-                self.prefix, 'The following error occurred: %s' % err))
-            KAFKA_ERRORS.inc()
-            self.should_run = False
-            print_error_loop_event()
+            LOG.error(
+                '%s - The following error occurred: %s',
+                prefix, err)
             raise KafkaMsgHandlerError(
-                format_message(
-                    self.prefix,
-                    'The following exception occurred: %s' % err,
-                    account_number=self.account_number,
-                    report_platform_id=self.report_platform_id))
-        finally:
-            await self.producer.stop()
+                '%s - The following exception occurred: %s for account=%s and report_platform_id=%s.',
+                prefix, err, self.account, self.report_platform_id)
